requirements.txt
attrs==20.3.0
Automat==20.2.0
backcall==0.1.0
bcrypt==3.2.0
bleach==3.3.0
canonicaljson==1.4.0
certifi==2020.4.5.1
cffi==1.14.5
chardet==3.0.4
constantly==15.1.0
cryptography==3.4.7
decorator==4.4.2
frozendict==1.2
hyperlink==21.0.0
idna==2.9
importlib-metadata==4.0.0
incremental==21.3.0
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
matrix-client==0.3.2
matrix-synapse==1.31.0
msgpack==1.0.2
netaddr==0.8.0
packaging==20.9
paho-mqtt==1.5.0
parso==0.7.0
pexpect==4.8.0
phonenumbers==8.12.21
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
PyYAML==5.4.1
requests==2.23.0
service-identity==18.1.0
signedjson==1.1.1
simplejson==3.17.2
six==1.14.0
socks==0
sortedcontainers==2.3.0
traitlets==4.3.3
treq==21.1.0
Twisted==21.2.0
typing-extensions==3.7.4.3
unpaddedbase64==2.1.0
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
zipp==3.4.1
zope.interface==5.4.0
